Flytt journalførDokument inn i JournalføringService og fiks tester som ikke kompilerer
