+++ src/main/kotlin/no/nav/familie/ba/sak/integrasjoner/journalføring/JournalføringService.kt
@@ -1,9 +1,11 @@
 package no.nav.familie.ba.sak.integrasjoner.journalføring
 
 import no.nav.familie.ba.sak.common.FunksjonellFeil
+import no.nav.familie.ba.sak.common.MDCOperations
 import no.nav.familie.ba.sak.ekstern.restDomene.RestFerdigstillOppgaveKnyttJournalpost
 import no.nav.familie.ba.sak.ekstern.restDomene.RestJournalføring
 import no.nav.familie.ba.sak.ekstern.restDomene.RestOppdaterJournalpost
+import no.nav.familie.ba.sak.integrasjoner.familieintegrasjoner.DEFAULT_JOURNALFØRENDE_ENHET
 import no.nav.familie.ba.sak.integrasjoner.familieintegrasjoner.IntegrasjonClient
 import no.nav.familie.ba.sak.integrasjoner.journalføring.domene.DbJournalpost
 import no.nav.familie.ba.sak.integrasjoner.journalføring.domene.DbJournalpostType
@@ -25,8 +27,14 @@ import no.nav.familie.ba.sak.kjerne.fagsak.FagsakService
 import no.nav.familie.ba.sak.kjerne.logg.LoggService
 import no.nav.familie.ba.sak.kjerne.steg.StegService
 import no.nav.familie.ba.sak.task.OpprettOppgaveTask
+import no.nav.familie.http.client.RessursException
 import no.nav.familie.kontrakter.felles.BrukerIdType
 import no.nav.familie.kontrakter.felles.Tema
+import no.nav.familie.kontrakter.felles.dokarkiv.Dokumenttype
+import no.nav.familie.kontrakter.felles.dokarkiv.v2.ArkiverDokumentRequest
+import no.nav.familie.kontrakter.felles.dokarkiv.v2.Dokument
+import no.nav.familie.kontrakter.felles.dokarkiv.v2.Filtype
+import no.nav.familie.kontrakter.felles.dokarkiv.v2.Førsteside
 import no.nav.familie.kontrakter.felles.journalpost.Bruker
 import no.nav.familie.kontrakter.felles.journalpost.Journalpost
 import no.nav.familie.kontrakter.felles.journalpost.JournalposterForBrukerRequest
@@ -34,6 +42,7 @@ import no.nav.familie.kontrakter.felles.journalpost.Journalstatus.FERDIGSTILT
 import no.nav.familie.kontrakter.felles.journalpost.Sak
 import no.nav.familie.kontrakter.felles.oppgave.Oppgavetype
 import org.slf4j.LoggerFactory
+import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Service
 import java.time.LocalDate
 import javax.transaction.Transactional
@@ -46,7 +55,7 @@ class JournalføringService(
     private val journalføringRepository: JournalføringRepository,
     private val loggService: LoggService,
     private val stegService: StegService,
-    private val journalføringMetrikk: JournalføringMetrikk
+    private val journalføringMetrikk: JournalføringMetrikk,
 ) {
 
     fun hentDokument(journalpostId: String, dokumentInfoId: String): ByteArray {
@@ -331,7 +340,93 @@ class JournalføringService(
         }
     }
 
+    fun journalførManueltBrev(
+        fnr: String,
+        fagsakId: String,
+        journalførendeEnhet: String,
+        brev: ByteArray,
+        dokumenttype: Dokumenttype,
+        førsteside: Førsteside?
+    ): String {
+        return journalførDokument(
+            fnr = fnr,
+            fagsakId = fagsakId,
+            journalførendeEnhet = journalførendeEnhet,
+            brev = listOf(
+                Dokument(
+                    dokument = brev,
+                    filtype = Filtype.PDFA,
+                    dokumenttype = dokumenttype
+                )
+            ),
+            førsteside = førsteside,
+        )
+    }
+
+    fun journalførDokument(
+        fnr: String,
+        fagsakId: String,
+        journalførendeEnhet: String? = null,
+        brev: List<Dokument>,
+        vedlegg: List<Dokument> = emptyList(),
+        førsteside: Førsteside? = null,
+        behandlingId: Long? = null
+    ): String {
+        if (journalførendeEnhet == DEFAULT_JOURNALFØRENDE_ENHET) {
+            logger.warn("Informasjon om enhet mangler på bruker og er satt til fallback-verdi, $DEFAULT_JOURNALFØRENDE_ENHET")
+        }
+
+        val eksternReferanseId = genererEksternReferanseIdForJournalpost(fagsakId, behandlingId)
+
+        val journalpostId = try {
+            val journalpost = integrasjonClient.journalførDokument(
+                ArkiverDokumentRequest(
+                    fnr = fnr,
+                    forsøkFerdigstill = true,
+                    hoveddokumentvarianter = brev,
+                    vedleggsdokumenter = vedlegg,
+                    fagsakId = fagsakId,
+                    journalførendeEnhet = journalførendeEnhet,
+                    førsteside = førsteside,
+                    eksternReferanseId = eksternReferanseId
+                )
+            )
+
+            if (!journalpost.ferdigstilt) {
+                error("Klarte ikke ferdigstille journalpost med id ${journalpost.journalpostId}")
+            }
+
+            journalpost.journalpostId
+        } catch (ressursException: RessursException) {
+            when (ressursException.httpStatus) {
+                HttpStatus.CONFLICT -> {
+                    logger.warn(
+                        "Klarte ikke journalføre dokument på fagsak=$fagsakId fordi det allerede finnes en journalpost " +
+                            "med eksternReferanseId=$eksternReferanseId. Bruker eksisterende journalpost."
+                    )
+
+                    hentEksisterendeJournalpost(eksternReferanseId, fnr)
+                }
+                else -> throw ressursException
+            }
+        }
+
+        return journalpostId
+    }
+
+    private fun hentEksisterendeJournalpost(
+        eksternReferanseId: String,
+        fnr: String
+    ): String = integrasjonClient.hentJournalposterForBruker(
+        JournalposterForBrukerRequest(
+            brukerId = Bruker(id = fnr, type = BrukerIdType.FNR),
+            antall = 50,
+        )
+    ).single { it.eksternReferanseId == eksternReferanseId }.journalpostId
+
     companion object {
+        fun genererEksternReferanseIdForJournalpost(fagsakId: String, behandlingId: Long?) =
+            "${fagsakId}_${behandlingId}_${MDCOperations.getCallId()}"
 
         private val logger = LoggerFactory.getLogger(JournalføringService::class.java)
     }
