+++ src/test/kotlin/no/nav/familie/ba/sak/kjerne/brev/DokumentServiceTest.kt
@@ -9,7 +9,7 @@ import no.nav.familie.ba.sak.config.AbstractSpringIntegrationTestDev
 import no.nav.familie.ba.sak.config.ClientMocks
 import no.nav.familie.ba.sak.config.DatabaseCleanupService
 import no.nav.familie.ba.sak.config.TEST_PDF
-import no.nav.familie.ba.sak.integrasjoner.familieintegrasjoner.IntegrasjonClient
+import no.nav.familie.ba.sak.integrasjoner.journalføring.JournalføringService
 import no.nav.familie.ba.sak.kjerne.arbeidsfordeling.ArbeidsfordelingService
 import no.nav.familie.ba.sak.kjerne.behandling.BehandlingHentOgPersisterService
 import no.nav.familie.ba.sak.kjerne.behandling.BehandlingService
@@ -68,7 +68,7 @@ class DokumentServiceTest(
     private val brevService: BrevService,
 
     @Autowired
-    private val integrasjonClient: IntegrasjonClient,
+    private val journalføringService: JournalføringService,
 
     @Autowired
     private val arbeidsfordelingService: ArbeidsfordelingService,
@@ -267,7 +267,7 @@ class DokumentServiceTest(
         dokumentService.sendManueltBrev(manueltBrevRequest, behandling, behandling.fagsak.id)
 
         io.mockk.verify(exactly = 1) {
-            integrasjonClient.journalførManueltBrev(
+            journalføringService.journalførManueltBrev(
                 fnr = manueltBrevRequest.mottakerIdent,
                 fagsakId = behandling.fagsak.id.toString(),
                 journalførendeEnhet = any(),
